apply_system_defaults: keep parameter empty if encryption fails
In apply_partnerId_default_values(), we iterate over all parameters and
if it's an encrypted one, we decrypt it. However, if decryption fails,
some undefined value is left in buf and this is copied into the value.
In practice, the previously decrypted value is reused.

Fix this by explicitly clearing the value if decryption fails. An empty
value is a better indication that something is wrong than some random
value.

Since I'm too lazy to copy this pattern four times, do the decryption in
a loop instead of copy&pasting the same code four times.