Revert "RDKSI-5819 : Address Review comments given by CMF/DT as part of upstreaming activity."
This reverts commit 69d7677105b5f9d2a79cc1990b5edc0002959dd3.

For RDKBDEV-706 we had one comment from CMF to convert hard coded macro to use syscfg var, (ra_interval)
so we have taken care of that now in our code but when we revisited the CMF review, we see that the older
change-set which had macro is already merged to comcast stable branch. So CMF is fine with old code itself.